refactor out code to seperate source files
